requirements.txt
matplotlib
tensorflow
scikit-learn
scikit-image
keras
gradio
python-dotenv